creating abstraction for date formatting
