+++ app/utils/date-fns.ts
@@ -1,3 +1,6 @@
+import type { ClientHint } from "~/modules/booking/types";
+import { getDateTimeFormatFromHints } from "./client-hints";
+
 export function getDifferenceInSeconds(
   dateLeft: Date,
   dateRight: Date
@@ -50,4 +53,28 @@ export function getTimeRemainingMessage(date1: Date, date2: Date): string {
   } else {
     return ""; //this should not happen
   }
+}
+
+export function formatDatesForICal(date: Date, hints: ClientHint) {
+  const dateTimeFormat = getDateTimeFormatFromHints(hints, {
+    year: "numeric",
+    month: "2-digit",
+    day: "2-digit",
+    hour: "2-digit",
+    minute: "2-digit",
+    second: "2-digit",
+    hour12: false,
+  });
+  const formatLocalDate = (date: Date, dateTimeFormat: Intl.DateTimeFormat) => {
+    const parts = dateTimeFormat.formatToParts(date);
+    const year = parts.find((part) => part.type === "year")!.value;
+    const month = parts.find((part) => part.type === "month")!.value;
+    const day = parts.find((part) => part.type === "day")!.value;
+    const hour = parts.find((part) => part.type === "hour")!.value;
+    const minute = parts.find((part) => part.type === "minute")!.value;
+    const second = parts.find((part) => part.type === "second")!.value;
+    return `${year}${month}${day}T${hour}${minute}${second}`;
+  };
+
+  return formatLocalDate(date, dateTimeFormat);
 }
